# docs/api/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
...
  /api/images/{id}/tweet-image4:
    get:
      tags:
      - rest-image-controller
      summary: Get the fourth media associated to a tweet
      operationId: downloadTweetImage4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "202":
          description: Fourth media not found
        "200":
          description: Fourth media obtained
          content:
            application/json:
              schema:
                type: object
  /api/images/{id}/tweet-image3:
    get:
      tags:
      - rest-image-controller
      summary: Get the third media associated to a tweet
      operationId: downloadTweetImage3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "202":
          description: Third media not found
        "200":
          description: Third media obtained
          content:
            application/json:
              schema:
                type: object
  /api/images/{id}/tweet-image2:
    get:
      tags:
      - rest-image-controller
      summary: Get the second media associated to a tweet
      operationId: downloadTweetImage2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Second media obtained
          content:
            application/json:
              schema:
                type: object
        "202":
          description: Second media not found
  /api/images/{id}/tweet-image1:
    get:
      tags:
      - rest-image-controller
      summary: Get the first media associated to a tweet
      operationId: downloadTweetImage1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "202":
          description: First media not found
        "200":
          description: First media obtained
          content:
            application/json:
              schema:
                type: object
  /api/images/{id}/profile-pic:
    get:
      tags:
      - rest-image-controller
      summary: Get the profile pic associated to a user
      operationId: downloadProfilePic
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Profile pic obtained
          content:
            application/json:
              schema:
                type: object
        "202":
          description: Profile pic not found
  /api/images/{id}/banner-pic:
    get:
      tags:
      - rest-image-controller
      summary: Get the banner pic associated to a user
      operationId: downloadBannerPic
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Banner pic obtained
          content:
            application/json:
              schema:
                type: object
        "202":
          description: Banner pic not found
...